PC-338 change component name
