+++ src/Assets/Elements/MedicalIssue/Style.js
