+++ src/Components/Generic/EncounterSheet/index.js
@@ -23,10 +23,6 @@ const EncounterSheet = ({
 }) => {
   const { t } = useTranslation();
 
-  const [listAllergy, setAllergyList] = useState([]);
-  const [listMedicalProblem, setMedicalProblem] = useState([]);
-  const [medicalProblemIsUpdated, setMedicalProblemIsUpdated] = useState(0); //for online update
-
   const isTabletMode = useMediaQuery(
     `(max-width: ${devicesValue.tabletPortrait}px)`,
   );
@@ -44,71 +40,6 @@ const EncounterSheet = ({
       url: '#',
     },
   ];
-
-  useEffect(() => {
-    (async () => {
-      try {
-        const listAllergyResult = await FHIR('Condition', 'doWork', {
-          functionName: 'getConditionListByParams',
-          functionParams: {
-            category: 'allergy',
-            subject: patient.id,
-            status: '1',
-          },
-        });
-        if (listAllergyResult.data && listAllergyResult.data.total > 0) {
-          let normalizedlistAllergy = [];
-          // eslint-disable-next-line
-          listAllergyResult.data.entry.map((res, id) => {
-            if (res.resource && res.resource.resourceType === 'Condition') {
-              let allergy = normalizeFhirCondition(res.resource);
-              normalizedlistAllergy.push(allergy);
-            }
-          });
-          setAllergyList(normalizedlistAllergy);
-        }
-      } catch (e) {
-        console.log('Error: ' + e);
-      }
-    })();
-
-    (async () => {
-      try {
-        const listMedicalProblemResult = await FHIR('Condition', 'doWork', {
-          functionName: 'getConditionListByParams',
-          functionParams: {
-            category: 'medical_problem',
-            subject: patient.id,
-            status: '1',
-          },
-        });
-        if (
-          listMedicalProblemResult.data &&
-          listMedicalProblemResult.data.total > 0
-        ) {
-          let normalizedListMedicalProblem = [];
-          // eslint-disable-next-line
-          listMedicalProblemResult.data.entry.map((res, id) => {
-            if (res.resource && res.resource.resourceType === 'Condition') {
-              let medicalProblem = normalizeFhirCondition(res.resource);
-              normalizedListMedicalProblem.push(medicalProblem);
-            }
-          });
-          setMedicalProblem(normalizedListMedicalProblem);
-        }
-      } catch (e) {
-        console.log('Error: ' + e);
-      }
-    })();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [medicalProblemIsUpdated]);
-
-  const callBackMedicalProblemChange = (listAllergy, listMedicalProblem) => {
-    // setMedicalProblemIsUpdated(Moment()); for online from db
-    setAllergyList(listAllergy);
-    setMedicalProblem(listMedicalProblem);
-  };
-
 
   const handleCloseClick = () => {
     history.push(`${firstRouteMapper(verticalName)}`);
@@ -133,15 +64,12 @@ const EncounterSheet = ({
           patient={patient}
           formatDate={formatDate}
           languageDirection={languageDirection}
-          listAllergy={listAllergy}
-          listMedicalProblem={listMedicalProblem}
         />
         <EncounterForms
           encounter={encounter}
           patient={patient}
           formatDate={formatDate}
           languageDirection={languageDirection}
-          changeMedicalProblem={callBackMedicalProblemChange}
         />
       </StyledEncounterSheet>
     </React.Fragment>
